rename: App.js 위치 root폴더로 위치 변경
